# 默认模块v1.0.0.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "答题系统接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/user/users/{id}": {
      "get": {
        "summary": "根据id查询用户",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/user/register": {
      "post": {
        "summary": "用户注册",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "userName",
                  "password"
                ]
              },
              "example": {
                "userName": "testuser",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                },
                "example": {
                  "code": 0,
                  "message": "success"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/user/login": {
      "post": {
        "summary": "用户登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "userName",
                  "password"
                ]
              },
              "example": {
                "userName": "testuser",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功，返回token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "token"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "token": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwidXNlck5hbWUiOiJ0ZXN0dXNlciIsInJvbGUiOjAsImlhdCI6MTc2NTY0MTI4NCwiZXhwIjoxNzY1NjQ4NDg0fQ.W8VyErMXxb746FS9zIMMqivO5yIerHoLZMy_PGvituc"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/user/info": {
      "get": {
        "summary": "获取当前用户信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer"
                        },
                        "userName": {
                          "type": "string"
                        },
                        "userRole": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "userId",
                        "userName",
                        "userRole"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "userId": 2,
                    "userName": "testuser",
                    "userRole": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/logout": {
      "post": {
        "summary": "退出登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                },
                "example": {
                  "code": 0,
                  "message": "success"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/quiz/questions": {
      "get": {
        "summary": "获取题目列表",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionId": {
                            "type": "integer"
                          },
                          "questionText": {
                            "type": "string"
                          },
                          "answer1Text": {
                            "type": "string"
                          },
                          "answer1Correct": {
                            "type": "integer"
                          },
                          "answer2Text": {
                            "type": "string"
                          },
                          "answer2Correct": {
                            "type": "integer"
                          },
                          "answer3Text": {
                            "type": "string"
                          },
                          "answer3Correct": {
                            "type": "integer"
                          },
                          "answer4Text": {
                            "type": "string"
                          },
                          "answer4Correct": {
                            "type": "integer"
                          },
                          "isDeleted": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "questionId",
                          "questionText",
                          "answer1Text",
                          "answer1Correct",
                          "answer2Text",
                          "answer2Correct",
                          "answer3Text",
                          "answer3Correct",
                          "answer4Text",
                          "answer4Correct",
                          "isDeleted",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": [
                    {
                      "questionId": 2,
                      "questionText": "以下哪一个是奇数？",
                      "answer1Text": "2",
                      "answer1Correct": 0,
                      "answer2Text": "3",
                      "answer2Correct": 1,
                      "answer3Text": "4",
                      "answer3Correct": 0,
                      "answer4Text": "6",
                      "answer4Correct": 0,
                      "isDeleted": 0,
                      "createdAt": "2025-12-13T22:38:09",
                      "updatedAt": "2025-12-13T22:38:09"
                    },
                    {
                      "questionId": 1,
                      "questionText": "中国的首都是哪座城市？",
                      "answer1Text": "北京",
                      "answer1Correct": 1,
                      "answer2Text": "上海",
                      "answer2Correct": 0,
                      "answer3Text": "广州",
                      "answer3Correct": 0,
                      "answer4Text": "深圳",
                      "answer4Correct": 0,
                      "isDeleted": 0,
                      "createdAt": "2025-12-13T22:38:09",
                      "updatedAt": "2025-12-13T22:38:09"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/quiz/questions/{id}": {
      "get": {
        "summary": "获取单个题目",
        "deprecated": false,
        "description": "",
        "tags": [
          "答题"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "questionId": {
                          "type": "integer"
                        },
                        "questionText": {
                          "type": "string"
                        },
                        "answer1Text": {
                          "type": "string"
                        },
                        "answer1Correct": {
                          "type": "integer"
                        },
                        "answer2Text": {
                          "type": "string"
                        },
                        "answer2Correct": {
                          "type": "integer"
                        },
                        "answer3Text": {
                          "type": "string"
                        },
                        "answer3Correct": {
                          "type": "integer"
                        },
                        "answer4Text": {
                          "type": "string"
                        },
                        "answer4Correct": {
                          "type": "integer"
                        },
                        "isDeleted": {
                          "type": "integer"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "questionId",
                        "questionText",
                        "answer1Text",
                        "answer1Correct",
                        "answer2Text",
                        "answer2Correct",
                        "answer3Text",
                        "answer3Correct",
                        "answer4Text",
                        "answer4Correct",
                        "isDeleted",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 0,
                  "message": "success",
                  "data": {
                    "questionId": 1,
                    "questionText": "中国的首都是哪座城市？",
                    "answer1Text": "北京",
                    "answer1Correct": 1,
                    "answer2Text": "上海",
                    "answer2Correct": 0,
                    "answer3Text": "广州",
                    "answer3Correct": 0,
                    "answer4Text": "深圳",
                    "answer4Correct": 0,
                    "isDeleted": 0,
                    "createdAt": "2025-12-13T22:38:09",
                    "updatedAt": "2025-12-13T22:38:09"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}